+++ .gitignore
@@ -1,3 +1,8 @@
+
+# Created by https://www.gitignore.io/api/csharp
+# Edit at https://www.gitignore.io/?templates=csharp
+
+### Csharp ###
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
 ##
@@ -29,7 +34,6 @@ bld/
 [Bb]in/
 [Oo]bj/
 [Ll]og/
-[Ll]ogs/
 
 # Visual Studio 2015/2017 cache/options directory
 .vs/
@@ -126,6 +130,9 @@ $tf/
 _ReSharper*/
 *.[Rr]e[Ss]harper
 *.DotSettings.user
+
+# JustCode is a .NET coding add-in
+.JustCode
 
 # TeamCity is a build add-in
 _TeamCity*
@@ -346,8 +353,9 @@ healthchecksdb
 # Backup folder for Package Reference Convert tool in Visual Studio 2017
 MigrationBackup/
 
-# Ionide (cross platform F# VS Code tools) working folder
-.ionide/
+# End of https://www.gitignore.io/api/csharp
 
-# Live unit testing files.
-*.lutconfig
+# Mac specific files
+.DS_Store
+
+*.sqlite
